Damage collision bug fixed. CameraShake improved
