+++ Assets/Scripts/Camera/CameraController.cs
@@ -8,6 +8,9 @@ public class CameraController : MonoBehaviour
     private Transform target;
     public float camSize = 10;
     public float cameraOffsetFactor = 0.25f;
+    
+    public float shakeIntensity = 0.5f; // Intensität der Kamerabewegung
+    public float shakeDuration = 0.02f; // Dauer der Kamerabewegung
 
     private Vector3 mousePosition;
 
@@ -33,5 +36,28 @@ public class CameraController : MonoBehaviour
     public void SetTarget(Transform t)
     {
         target = t;
+    }
+    
+    public void StartShaking(Vector2 shotDirection)
+    {
+        StartCoroutine(ShakeCamera(shotDirection));
+    }
+    private IEnumerator ShakeCamera(Vector2 shotDirection)
+    {
+        Vector3 originalPosition = transform.localPosition; // Ursprüngliche Position der Kamera
+
+        float elapsedTime = 0f;
+
+        while (elapsedTime < shakeDuration)
+        {
+            // Berechne die Verschiebung basierend auf der Schussrichtung und Intensität
+            Vector2 offset = -shotDirection * shakeIntensity;
+
+            transform.localPosition = originalPosition + new Vector3(offset.x, offset.y, 0f);
+            elapsedTime += Time.deltaTime;
+            yield return null;
+        }
+
+        transform.localPosition = originalPosition; // Zurücksetzen der Kameraposition
     }
 }
