--- Assets/Scripts/Camera/CameraShake.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: da0ac78b50dc074469afc136586f1c6b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
