+++ UserSettings/EditorUserSettings.asset
@@ -29,6 +29,9 @@ EditorUserSettings:
     RecentlyUsedSceneGuid-7:
       value: 0555525200575c5d555a0d7542220d44444e1d287c2e776379714e32b2b4616d
       flags: 0
+    RecentlyUsedSceneGuid-8:
+      value: 0604505f5c000f5a0f5b5f74167b0b44174e4f72752d27637a7b4b66b0e16168
+      flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
       flags: 0
